Move shared KDA host factory utilities
